requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.10
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.6.2.post1
argcomplete==3.6.2
attrs==25.3.0
Authlib==1.6.0
bcrypt==3.2.0
beautifulsoup4==4.13.4
boto3==1.38.29
botocore==1.38.29
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cohere==5.15.0
colorama==0.4.6
cryptography==43.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fasta2a==0.2.14
fastapi==0.115.2
fastavro==1.11.1
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
google==3.0.0
google-adk==1.1.1
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.96.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.18.0
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.7.3
groq==0.26.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.14.0
hf-xet==1.1.3
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.6.1
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.45
logfire-api==3.17.0
markdown-it-py==3.0.0
marshmallow==3.26.1
mcp==1.9.2
mdurl==0.1.2
mistralai==1.8.1
motor==3.6.0
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.84.0
openpyxl==3.1.5
opentelemetry-api==1.33.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
passlib==1.7.4
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic-ai==0.2.14
pydantic-ai-slim==0.2.14
pydantic-evals==0.2.14
pydantic-graph==0.2.14
pydantic-settings==2.9.1
pydantic_core==2.23.4
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.9.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
serpapi==0.1.5
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.3.6
starlette==0.40.0
tenacity==9.1.2
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250602
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
redis==5.2.0
celery==5.4.0
